routes-config: Rename as routesConfig
Rename file as routesConfig in order to be in line with the other file naming conventions.
Also update the import statement in index.ts.
